+++ resources/js/controllers/fields/tinymce_controller.js
@@ -65,7 +65,7 @@ export default class extends Controller {
         //    /.*/
         // );
 
-        tinymce.baseURL = '/orchid/js/tinymce';
+        tinymce.baseURL = '/orchid/orchid/js/tinymce';
 
         const selector = this.element.querySelector('.tinymce').id;
         const input = this.element.querySelector('input');