+++ resources/views/layouts/app.blade.php
@@ -7,7 +7,7 @@
     <title>@yield('title','ORCHID') - @yield('description','Admin')</title>
     <meta name="csrf_token" content="{{csrf_token()}}">
     <meta name="auth" content="{{Auth::check()}}">
-    <link rel="stylesheet" type="text/css" href="{{mix('/css/orchid.css','orchid')}}">
+    <link rel="stylesheet" type="text/css" href="{{orchid_mix('/css/orchid.css','orchid')}}">
 
     @stack('head')
 
@@ -17,9 +17,9 @@
     <meta http-equiv="X-DNS-Prefetch-Control" content="on"/>
     <link rel="dns-prefetch" href="{{ config('app.url') }}"/>
 
-    <script src="{{ mix('/js/manifest.js','orchid')}}" type="text/javascript"></script>
-    <script src="{{ mix('/js/vendor.js','orchid')}}" type="text/javascript"></script>
-    <script src="{{ mix('/js/orchid.js','orchid')}}" type="text/javascript"></script>
+    <script src="{{ orchid_mix('/js/manifest.js','orchid')}}" type="text/javascript"></script>
+    <script src="{{ orchid_mix('/js/vendor.js','orchid')}}" type="text/javascript"></script>
+    <script src="{{ orchid_mix('/js/orchid.js','orchid')}}" type="text/javascript"></script>
 
     @foreach(Dashboard::getResource('stylesheets') as $stylesheet)
         <link rel="stylesheet" href="{{$stylesheet}}">