+++ routes/dashboard.php
@@ -2,6 +2,7 @@
 
 declare(strict_types=1);
 
+use Orchid\Platform\Http\Controllers\Systems\IndexController;
 use Orchid\Platform\Http\Screens\BackupScreen;
 use Orchid\Platform\Http\Screens\AnnouncementScreen;
 use Orchid\Platform\Http\Controllers\Systems\SearchController;
@@ -16,13 +17,8 @@ use Orchid\Platform\Http\Controllers\Systems\SearchController;
 */
 
 // Index and default...
-$this->get('/', function () {
-    return redirect()->route(config('platform.index'));
-})->name('platform.index');
-
-$this->fallback(function () {
-    return view('platform::errors.404');
-});
+$this->get('/', [IndexController::class, 'index'])->name('platform.index');
+$this->fallback([IndexController::class,'fallback']);
 
 $this->post('/search/{query}', [SearchController::class, 'index'])->name('platform.search');