+++ tests/Environment.php
@@ -46,7 +46,6 @@ trait Environment
         $this->loadLaravelMigrations();
         $this->loadMigrationsFrom(realpath('./database/migrations'));
         $this->artisan('migrate', ['--database' => 'orchid']);
-        $this->artisan('orchid:link');
 
         $this->withFactories(realpath(PLATFORM_PATH.'/database/factories'));